+++ instances/herrenberg/android/app/src/main/kotlin/de/stadtnavi/herrenberg/MainActivity.kt
@@ -1,4 +1,4 @@
-package de.stadtnavi.herrenberg
+package de.stadtnavi.herrenberg.internal
 
 import io.flutter.embedding.android.FlutterActivity