+++ lib/base/custom_layers/live_bus/live_bus_layer.dart
@@ -121,6 +121,16 @@ class LiveBusLayer extends CustomLayer {
   }
 
   @override
+  List<Marker>? buildLayerMarkersPriority(int? zoom) {
+    return [];
+  }
+
+  @override
+  Widget? buildLayerOptionsBackground(int? zoom) {
+    return null;
+  }
+
+  @override
   Widget buildLayerOptions(int? zoom) {
     double? markerSize;
     switch (zoom) {