+++ lib/base/custom_layers/pbf_layer/bicycle_network/bicycle_network_layer.dart
@@ -1,7 +1,9 @@
 import 'dart:convert';
 
 import 'package:flutter/material.dart';
+import 'package:flutter_map_marker_cluster/flutter_map_marker_cluster.dart';
 import 'package:http/http.dart' as http;
+import 'package:flutter_map/flutter_map.dart';
 import 'package:latlong2/latlong.dart';
 
 import 'package:stadtnavi_core/base/custom_layers/custom_layer.dart';
@@ -79,7 +81,12 @@ class BicycleNetworkLayer extends CustomLayer {
   }
 
   @override
-  Widget buildLayerOptions(int? zoom) {
+  List<Marker>? buildLayerMarkersPriority(int? zoom) {
+    return [];
+  }
+
+  @override
+  Widget? buildLayerOptionsBackground(int? zoom) {
     if (_listlist.isEmpty) {
       load();
     }
@@ -114,6 +121,11 @@ class BicycleNetworkLayer extends CustomLayer {
               .toList()
           : [],
     );
+  }
+
+  @override
+  Widget buildLayerOptions(int? zoom) {
+    return PolygonLayer();
   }
 
   @override