+++ lib/base/custom_layers/pbf_layer/charging/charging_layer.dart
@@ -6,6 +6,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:http/http.dart' as http;
 import 'package:stadtnavi_core/base/custom_layers/cubits/panel/panel_cubit.dart';
 import 'package:stadtnavi_core/base/custom_layers/custom_layer.dart';
+import 'package:stadtnavi_core/base/custom_layers/models/enums.dart';
 import 'package:stadtnavi_core/base/custom_layers/static_layer.dart';
 import 'package:stadtnavi_core/consts.dart';
 import 'package:trufi_core/base/translations/trufi_base_localizations.dart';
@@ -18,12 +19,98 @@ import 'charging_marker_modal.dart';
 class ChargingLayer extends CustomLayer {
   final Map<String, ChargingFeature> _pbfMarkers = {};
 
+  Map<String, ChargingFeature> get data => _pbfMarkers;
+
   ChargingLayer(String id, String weight) : super(id, weight);
   void addMarker(ChargingFeature pointFeature) {
     if (_pbfMarkers[pointFeature.id] == null) {
       _pbfMarkers[pointFeature.id] = pointFeature;
       refresh();
     }
+  }
+
+  @override
+  List<Marker>? buildLayerMarkersPriority(int? zoom) {
+    double? markerSize;
+    switch (zoom) {
+      case 15:
+        markerSize = 20;
+        break;
+      case 16:
+        markerSize = 25;
+        break;
+      case 17:
+        markerSize = 30;
+        break;
+      case 18:
+        markerSize = 35;
+        break;
+      default:
+        markerSize = zoom != null && zoom > 18 ? 40 : null;
+    }
+    final markersList = _pbfMarkers.values.toList();
+    // avoid vertical wrong overlapping
+    markersList.sort(
+      (b, a) => a.position.latitude.compareTo(b.position.latitude),
+    );
+    return markerSize != null
+        ? markersList
+            .map((element) => Marker(
+                  key: Key("$id:${element.id}"),
+                  height: markerSize!,
+                  width: markerSize,
+                  point: element.position,
+                  anchorPos: AnchorPos.align(AnchorAlign.center),
+                  builder: (context) {
+                    final availabilityStatus = element.getAvailabilityStatus();
+                    return GestureDetector(
+                      onTap: () {
+                        final panelCubit = context.read<PanelCubit>();
+                        panelCubit.setPanel(
+                          CustomMarkerPanel(
+                            panel: (
+                              context,
+                              onFetchPlan, {
+                              isOnlyDestination,
+                            }) =>
+                                ChargingMarkerModal(
+                              element: element,
+                              onFetchPlan: onFetchPlan,
+                            ),
+                            positon: element.position,
+                            minSize: 50,
+                          ),
+                        );
+                      },
+                      child: Stack(
+                        children: [
+                          Container(
+                            margin: EdgeInsets.only(
+                              left: markerSize! / 5,
+                              top: markerSize / 5,
+                            ),
+                            child: SvgPicture.string(
+                              chargingIcon,
+                            ),
+                          ),
+                          if (availabilityStatus != null)
+                            SizedBox(
+                              height: markerSize / 1.8,
+                              width: markerSize / 1.8,
+                              child: availabilityStatus.getImage(),
+                            ),
+                        ],
+                      ),
+                    );
+                  },
+                ))
+            .toList()
+        : [];
+  }
+
+  @override
+  Widget? buildLayerOptionsBackground(int? zoom) {
+    return null;
   }
 
   @override
@@ -58,84 +145,49 @@ class ChargingLayer extends CustomLayer {
                     width: markerSize,
                     point: element.position,
                     anchorPos: AnchorPos.align(AnchorAlign.center),
-                    builder: (context) => GestureDetector(
-                      onTap: () {
-                        final panelCubit = context.read<PanelCubit>();
-                        panelCubit.setPanel(
-                          CustomMarkerPanel(
-                            panel: (
-                              context,
-                              onFetchPlan, {
-                              isOnlyDestination,
-                            }) =>
-                                ChargingMarkerModal(
-                              element: element,
-                              onFetchPlan: onFetchPlan,
+                    builder: (context) {
+                      final availabilityStatus =
+                          element.getAvailabilityStatus();
+                      return GestureDetector(
+                        onTap: () {
+                          final panelCubit = context.read<PanelCubit>();
+                          panelCubit.setPanel(
+                            CustomMarkerPanel(
+                              panel: (
+                                context,
+                                onFetchPlan, {
+                                isOnlyDestination,
+                              }) =>
+                                  ChargingMarkerModal(
+                                element: element,
+                                onFetchPlan: onFetchPlan,
+                              ),
+                              positon: element.position,
+                              minSize: 50,
                             ),
-                            positon: element.position,
-                            minSize: 50,
-                          ),
-                        );
-                      },
-                      child: Stack(
-                        children: [
-                          Container(
-                            margin: EdgeInsets.only(
-                              left: markerSize! / 5,
-                              top: markerSize / 5,
+                          );
+                        },
+                        child: Stack(
+                          children: [
+                            Container(
+                              margin: EdgeInsets.only(
+                                left: markerSize! / 5,
+                                top: markerSize / 5,
+                              ),
+                              child: SvgPicture.string(
+                                chargingIcon,
+                              ),
                             ),
-                            child: SvgPicture.string(
-                              chargingIcon,
-                            ),
-                          ),
-                          if (element.capacityUnknown == 0)
-                            if ((element.available ?? 0) > 1)
-                              Positioned(
-                                child: Container(
-                                  height: markerSize / 1.8,
-                                  width: markerSize / 1.8,
-                                  decoration: BoxDecoration(
-                                    color: Colors.green,
-                                    borderRadius: BorderRadius.circular(100),
-                                  ),
-                                  child: Center(
-                                    child: Icon(
-                                      Icons.check,
-                                      size: markerSize / 3,
-                                      color: Colors.white,
-                                    ),
-                                  ),
-                                ),
-                              )
-                            else if ((element.available ?? 0) == 1)
-                              Positioned(
-                                child: SizedBox(
-                                  height: markerSize / 1.8,
-                                  width: markerSize / 1.8,
-                                  child: SvgPicture.string(
-                                    charginIconPoorAvailability,
-                                  ),
-                                ),
-                              )
-                            else
-                              Container(
+                            if (availabilityStatus != null)
+                              SizedBox(
                                 height: markerSize / 1.8,
                                 width: markerSize / 1.8,
-                                decoration: BoxDecoration(
-                                  color: Colors.red,
-                                  borderRadius: BorderRadius.circular(100),
-                                ),
-                                child: Center(
-                                  child: Icon(
-                                    Icons.close,
-                                    size: markerSize / 3,
-                                    color: Colors.white,
-                                  ),
-                                ),
-                              )
-                        ],
-                      ),
-                    ),
+                                child: availabilityStatus.getImage(),
+                              ),
+                          ],
+                        ),
+                      );
+                    },
                   ))
               .toList()
           : zoom != null && zoom > 11