+++ lib/base/custom_layers/static_layer.dart
@@ -13,20 +13,20 @@ import 'pbf_layer/bike_parks/bike_parks_layer.dart';
 
 class StaticTileLayers {
   static Map<StopsLayerIds, StopsLayer> stopsLayers = {
-    StopsLayerIds.bus: StopsLayer(StopsLayerIds.bus, '3'),
-    StopsLayerIds.rail: StopsLayer(StopsLayerIds.rail, '3'),
-    StopsLayerIds.carpool: StopsLayer(StopsLayerIds.carpool, '3'),
-    StopsLayerIds.subway: StopsLayer(StopsLayerIds.subway, '3'),
+    StopsLayerIds.bus: StopsLayer(StopsLayerIds.bus, '4'),
+    StopsLayerIds.rail: StopsLayer(StopsLayerIds.rail, '4'),
+    StopsLayerIds.carpool: StopsLayer(StopsLayerIds.carpool, '4'),
+    StopsLayerIds.subway: StopsLayer(StopsLayerIds.subway, '4'),
   };
-  static ParkingLayer parkingLayer = ParkingLayer("Parking", '3');
+  static ParkingLayer parkingLayer = ParkingLayer("Parking", '4');
   static ParkingZonesLayer parkingZonesLayer =
       ParkingZonesLayer("Parking Zones", '1');
-  static CityBikesLayer citybikeLayer = CityBikesLayer("Sharing", '3');
-  static BikeParkLayer bikeParkLayer = BikeParkLayer("Bike Parking Space", '3');
+  static CityBikesLayer citybikeLayer = CityBikesLayer("Sharing", '5');
+  static BikeParkLayer bikeParkLayer = BikeParkLayer("Bike Parking Space", '4');
   static BicycleNetworkLayer bicycleNetworkLayer =
       BicycleNetworkLayer("Bicycle Network Space", '2');
-  static CifsLayer cifsLayer = CifsLayer("Roadworks", '2');
+  static CifsLayer cifsLayer = CifsLayer("Roadworks", '3');
   static LiveBusLayer liveBusLayer = LiveBusLayer("LiveBusBeta", '3');
   static WeatherLayer weatherLayer = WeatherLayer("Road Weather", '3');
-  static ChargingLayer chargingLayer = ChargingLayer("Charging", '3');
+  static ChargingLayer chargingLayer = ChargingLayer("Charging", '4');
 }