+++ lib/base/pages/home/home_page.dart
@@ -258,8 +258,10 @@ class _HomePageState extends State<HomePage>
     }
     widget.asyncExecutor.run(
       context: context,
-      onExecute: () async =>
-          await mapRouteCubit.fetchPlan(advancedOptions: settingFetchState),
+      onExecute: () async {
+        await mapModesCubit.reset();
+        await mapRouteCubit.fetchPlan(advancedOptions: settingFetchState);
+      },
       onFinish: (_) {
         mapModesCubit.fetchModesPlans(
           from: mapRouteState.fromPlace!,