+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## 0.0.1
+## 1.7.0
+🎉 Support for Flutter 3.3.3
+🎉 Support trufi-core_v3.2.0-stadtnavi-lts
+🎉 Overlapping icons
+🎉 Improve POI's Layers
 
-* TODO: Describe initial release.
+## 1.6.0
+🎉 Support for Flutter 3.3.3
+🎉 Support trufi-core_v3.2.0-stadtnavi-lts
+🎉 Push notifications
+🎉 Improve features